refactor: make TimeSpan great again
We had a few asserts that a timespan tuple has its end after its start,
with the TODO to replace them. We have now replaced them.
Introduced a new type to ensure the invariant that a time span is valid
(start < end) and also to have a few convenience methods on it.